+++ anchore_engine/services/__init__.py
@@ -1,5 +1,5 @@
 from .analyzer.service import AnalyzerService
 from .apiext import ExternalApiService
-from .catalog import CatalogService
+from .catalog.service import CatalogService
 from .policy_engine import PolicyEngineService
 from .simplequeue import SimpleQueueService
