+++ anchore_manager/cli/__init__.py
@@ -1,7 +1,8 @@
 import click
 
 from anchore_manager import version
-from anchore_manager.util.config import init_all
+from anchore_manager.util import logging
+from anchore_manager.util.config import setup_config
 
 from . import analyzers, db, objectstorage, service
 
@@ -16,8 +17,8 @@ from . import analyzers, db, objectstorage, service
 @click.pass_context
 def main_entry(ctx, debug, json, configdir):
     cli_opts = {"json": json, "debug": debug, "configdir": configdir}
-
-    config = init_all(cli_opts)
+    config = setup_config(cli_opts)
+    logging.log_config(config)
     ctx.obj = config
