+++ twisted/plugins/anchore_catalog.py
@@ -2,7 +2,7 @@ from twisted.application.service import IServiceMaker
 from twisted.plugin import IPlugin
 from zope.interface import implementer
 
-from anchore_engine.services.catalog import CatalogService
+from anchore_engine.services.catalog.service import CatalogService
 from anchore_engine.twisted import CommonOptions, WsgiApiServiceMaker
